refactor: types 파일 생성 및 분리
