+++ src/components/common/Modal/card-confirm/CardConfirmModal.tsx
@@ -1,6 +1,7 @@
 import MainBox from './MainBox';
 import ModalHeader from './ModalHeader';
 import SideBox from './SideBox';
+import { CardConfirmModalProps } from './types';
 import styled from 'styled-components';
 
 const S = {
@@ -17,28 +18,6 @@ const S = {
   `,
 };
 
-export interface Assignee {
-  profileImageUrl: string;
-  nickname: string;
-  id: number;
-}
-
-export interface CardInfoProps {
-  id: number;
-  title: string;
-  description: string;
-  tags: string[];
-  dueDate: string;
-  assignee: Assignee;
-  imageUrl: string;
-  teamId: string;
-  columnId: number;
-  createdAt: string;
-  updatedAt: string;
-}
-export interface CardConfirmModalProps {
-  cardInfoData: CardInfoProps;
-}
 function CardConfirmModal({ cardInfoData }: CardConfirmModalProps) {
   return (
     <S.ModalContainer>
