+++ src/components/common/Modal/card-confirm/types.ts
@@ -1,0 +1,22 @@
+export interface Assignee {
+  profileImageUrl: string;
+  nickname: string;
+  id: number;
+}
+
+export interface CardInfoProps {
+  id: number;
+  title: string;
+  description: string;
+  tags: string[];
+  dueDate: string;
+  assignee: Assignee;
+  imageUrl: string;
+  teamId: string;
+  columnId: number;
+  createdAt: string;
+  updatedAt: string;
+}
+export interface CardConfirmModalProps {
+  cardInfoData: CardInfoProps;
+}
